TS-46833 Extract the test arguments that all commands share
The two argument builders had drifted into holding the same connection
options twice, down to verbatim copies of getAccessKeyFromCi() and the
test user constant. CommonUploadArguments now holds everything
CommonCommandLineOptions registers -- server, project, user, access key,
proxy, keystore, stacktrace, debug, timeout and retries -- and emits them
via addCommonOptions(). The builders keep only what their own command
takes.

The setters return the concrete subclass so that they can still be
chained with the command-specific ones, which is what the self type on
CommonUploadArguments is for. AssertJ uses the same idiom.

getStdinFile() is no longer a default method on UploadArguments: every
builder inherits it from CommonUploadArguments, so nothing was left to
default to. The one caller that implemented the interface as a lambda,
and therefore relied on it having a single abstract method, now uses
TeamscaleUploadArguments.withHelp() instead. That mirrors the SBOM
builder and keeps the two help tests symmetric.

Common options are now emitted as a block, so the order of the assembled
command line changed. No test depends on it.
